Refactored Day 6 to new format
