+++ .github/dependabot.yml
@@ -148,16 +148,3 @@ updates:
       interval: daily
       time: "21:00"
     open-pull-requests-limit: 99
-  # 2019 - Day 6
-  - package-ecosystem: nuget
-    directory: "/2019/Day6/Solution"
-    schedule:
-      interval: daily
-      time: "21:00"
-    open-pull-requests-limit: 99
-  - package-ecosystem: nuget
-    directory: "/2019/Day6/Tests"
-    schedule:
-      interval: daily
-      time: "21:00"
-    open-pull-requests-limit: 99
