--- 2019/Day6/Solution/Day6.csproj
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>e27fdcf7-d252-4000-a18d-986c51dbd22f</UserSecretsId>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NAoCHelper" Version="1.2.5" />
-  </ItemGroup>
-</Project>
