--- 2019/Day6/Solution/appsettings.json
@@ -1,5 +1,0 @@
-{
-    "Secrets": {
-        "Cookie": "Configured in User Secrets"
-    }
-}
